Basic driver implementation for OpenRazer

+++ src/main/java/io/github/TheSuperGamer20578/Chroma/drivers/openRazer/DBusDeviceMisc.java
@@ -33,7 +33,7 @@ interface DBusDeviceMisc extends DBusInterface {
     /**
      * If the device has an LED matrix
      */
-    byte[] getMatrixDimensions();
+    int[] getMatrixDimensions();
 
     /**
      * If the device has an LED matrix
